requirements.txt
flask
flask-cors
flask-sqlalchemy
pyjwt==2.8.0
werkzeug
gunicorn
psycopg2-binary